{{cookiecutter.project_name}}/core/serializers/__init__.py
from rest_framework.serializers import ModelSerializer as DRFModelSerializer

from .abstract_base_serializer import AbstractBaseSerializer


class ModelSerializer(AbstractBaseSerializer, DRFModelSerializer):
    pass
