{{cookiecutter.project_name}}/requirements/base.txt
Django==4.0.7
django-environ==0.9.0
psycopg2==2.9.3
djangorestframework==3.13.1

celery==5.2.7
django-celery-results==2.4.0
django-celery-beat==2.3.0

django-filter==22.1
django-extensions==3.2.1
django-cors-headers==3.13.0

boto3==1.24.71
attrs==22.1.0

drf-yasg==1.21.5

django-redis==5.2.0

Faker==15.1.1
factory-boy==3.2.1
model-bakery==1.10.1

django-compressor==4.1

django-jalali-date==1.0.1
